Subpixel positioning

+++ ffng4/src/main/cpp/game/graphics/subtitles.cpp
@@ -80,7 +80,7 @@ void Subtitles::draw(const DrawTarget &target, float dTime, float absTime) {
             glUniform1f(textProgram.uniform("uTime"), absTime - line.addTime);
             auto height = line.texture.height();
             float destY = pixelSize.fy() - (float)height * (1.f + line.yOffset);
-            target.blit(line.texture, textProgram, 0, (int)destY - (int)height, 0, 0,
+            target.blit(line.texture, textProgram, 0, destY - (float)height, 0, 0,
                         DrawTarget::fullSize, 3 * height, pixelSize.x(), pixelSize.y());
         }
 }
